requirements.txt

# Core
Django>=4.2,<5

# API
djangorestframework>=3.14

# Environment/configuration
django-environ>=0.12.0
python-dotenv>=1.0.0

# Production / deployment helpers
gunicorn>=20.1.0
whitenoise>=6.5.0

# Database (Postgres adapter; optional if you only use sqlite)
psycopg2-binary>=2.9

# Frontend integration / CORS
django-cors-headers>=3.15

# Testing
pytest>=7.0
pytest-django>=4.5

pre-commit
black
ruff
isort

# Pin additional useful utilities as needed
# e.g. django-environ will parse DATABASE_URL environment variable
